+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopobjects/sign/SKSignShopObjectType.java
@@ -45,11 +45,11 @@ public final class SKSignShopObjectType extends BaseBlockShopObjectType<SKSignSh
 
 	@Override
 	public boolean validateSpawnLocation(
-			@Nullable Player creator,
+			@Nullable Player player,
 			@Nullable Location spawnLocation,
 			@Nullable BlockFace attachedBlockFace
 	) {
-		if (!super.validateSpawnLocation(creator, spawnLocation, attachedBlockFace)) {
+		if (!super.validateSpawnLocation(player, spawnLocation, attachedBlockFace)) {
 			return false;
 		}
 		assert spawnLocation != null;
@@ -57,8 +57,8 @@ public final class SKSignShopObjectType extends BaseBlockShopObjectType<SKSignSh
 		// Block has to be empty:
 		Block spawnBlock = spawnLocation.getBlock();
 		if (!spawnBlock.isEmpty()) {
-			if (creator != null) {
-				TextUtils.sendMessage(creator, Messages.spawnBlockNotEmpty);
+			if (player != null) {
+				TextUtils.sendMessage(player, Messages.spawnBlockNotEmpty);
 			}
 			return false;
 		}
@@ -67,8 +67,8 @@ public final class SKSignShopObjectType extends BaseBlockShopObjectType<SKSignSh
 		if (attachedBlockFace == BlockFace.DOWN
 				|| (attachedBlockFace == BlockFace.UP && !Settings.enableSignPostShops)
 				|| (attachedBlockFace != null && !BlockFaceUtils.isBlockSide(attachedBlockFace))) {
-			if (creator != null) {
-				TextUtils.sendMessage(creator, Messages.invalidSpawnBlockFace);
+			if (player != null) {
+				TextUtils.sendMessage(player, Messages.invalidSpawnBlockFace);
 			}
 			return false;
 		}